Extracting interfaces and common functions
